Patterns working
Patterns based on "x" number of squares can now be defined.

Patterns are defined by defining each square in the format (x,y,w,h), where x and y are the coordinates of the bottom left hand corner, w and h are the width and height, respectively. All units in m.

Example:
squares = [-1,-1,-1,1;1,1,1,1;-1,1,1,1;1,-1,1,1];

4 squares, each 1 m wide and high, corners at -1,-1; 1,1; -1,1 and 1,-1.